Add hook to fill like button when user likes movie
