+++ pages/reserve/[movieId].tsx
@@ -4,7 +4,6 @@ import { useQuery } from "react-query";
 import { Flex } from "@chakra-ui/react";
 import { useRouter } from "next/router";
 import { useSetRecoilState } from "recoil";
-import { doc, getDoc } from "firebase/firestore";
 
 import {
   BASE_QUERY,
@@ -12,18 +11,13 @@ import {
   fetchCredit,
   fetchDetail,
 } from "../../src/hooks/fetching";
-import {
-  ICast,
-  IMovieDetails,
-  IUserDB,
-  IUserMovie,
-} from "../../src/interfaces";
+import { ICast, IMovieDetails } from "../../src/interfaces";
 import Navigation from "../../src/components/Navigation/Navigation";
 import MovieDetail from "../../src/components/MovieDetail";
 import ShowTime from "../../src/components/ShowTime";
 import LoadingAnimation from "../../src/components/LoadingAnimation";
-import { likedState, movieIDState } from "../../src/atom";
-import { auth, db } from "../../firebase";
+import { movieIDState } from "../../src/atom";
+import useFillButton from "../../src/hooks/useFillButton";
 
 interface IReserveProps {
   isPlaying: boolean;
@@ -35,37 +29,13 @@ const Reserve: NextPage<IReserveProps> = ({ isPlaying }) => {
   } = useRouter(); // string
 
   const setMovieID = useSetRecoilState(movieIDState);
-  const setLiked = useSetRecoilState(likedState);
-
-  // 사용자가 찜한 영화 가져오기
-  const getLikedMovies = async (uid: string) => {
-    const docRef = doc(db, "users", uid);
-    const docSnap = await getDoc(docRef);
-    const movies: IUserDB["movies"] = docSnap.data()?.movies;
-
-    return movies;
-  };
 
   useEffect(() => {
     const id = parseInt(movieId as any, 10);
     setMovieID(id);
-
-    // 사용자 로그인 여부 확인
-    (async () => {
-      const user = auth.currentUser;
-      if (user) {
-        const { uid } = user;
-        const likedMovies = await getLikedMovies(uid); // 사용자가 찜한 영화
-        if (likedMovies !== undefined) {
-          const index = likedMovies.findIndex(
-            (movie: IUserMovie) => `${movie.id}` === movieId
-          );
-          // 좋아요 표시
-          setLiked(index === -1 ? false : true);
-        }
-      }
-    })();
   }, [movieId]);
+
+  useFillButton(movieId);
 
   const { data: detailData, isLoading: detailLoading } =
     useQuery<IMovieDetails>(["detail", movieId], () => fetchDetail(movieId));
